test dao fatti mancano da aggiungere gli isPart a group

+++ test/tests/ORMTest/FieldDaoTest.java
@@ -31,11 +31,8 @@ class FieldDaoTest extends GeneralDAOTest{
         String name = "Padel new sport";
         sport = createSport(name);
         facility = field.getFacility();
-        if (fieldDao.getField(field.getId()) == null)
+        if (field.getId() == 0)
             shouldSkip = true;
-
-
-        Assumptions.assumeFalse(shouldSkip);
     }
 
     @Override
@@ -53,12 +50,10 @@ class FieldDaoTest extends GeneralDAOTest{
         sportDao.deleteSport(sport.getId());
         sportDao.deleteSport(field.getSport().getId());
         facilityDao.deleteFacility(facility.getId());
-        ownerDao.deletePerson(facility.getOwner().getUsername());
-        if (!(fieldDao.getField(field.getId()) == null))
-            shouldSkip = true;
+        if (facility.getOwner() != null && facility.getOwner().getId() != 0)
+            ownerDao.deletePerson(facility.getOwner().getUsername());
 
-
-        Assumptions.assumeFalse(shouldSkip);
+        shouldSkip = false;
     }
 
     @Test
@@ -67,6 +62,8 @@ class FieldDaoTest extends GeneralDAOTest{
 
     @Test
     void getField() throws SQLException, ClassNotFoundException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertNotNull(fieldDao.getField(field.getId()));
     }
 
@@ -76,6 +73,8 @@ class FieldDaoTest extends GeneralDAOTest{
 
     @Test
     void updateName() throws SQLException, ClassNotFoundException {
+        Assumptions.assumeFalse(shouldSkip);
+
         String name = "Name new";
         fieldDao.updateName(field.getId(), name);
         assertEquals(name, fieldDao.getField(field.getId()).getName());
@@ -83,6 +82,8 @@ class FieldDaoTest extends GeneralDAOTest{
 
     @Test
     void updateDescription() throws SQLException, ClassNotFoundException {
+        Assumptions.assumeFalse(shouldSkip);
+
         String description = "Description new";
         fieldDao.updateDescription(field.getId(), description);
         assertEquals(description, fieldDao.getField(field.getId()).getDescription());
@@ -90,6 +91,8 @@ class FieldDaoTest extends GeneralDAOTest{
 
     @Test
     void updatePrice() throws SQLException, ClassNotFoundException {
+        Assumptions.assumeFalse(shouldSkip);
+
         int price = 99;
         fieldDao.updatePrice(field.getId(), price);
         assertEquals(price, fieldDao.getField(field.getId()).getPrice());
@@ -97,6 +100,8 @@ class FieldDaoTest extends GeneralDAOTest{
 
     @Test
     void updateImage() throws SQLException, ClassNotFoundException {
+        Assumptions.assumeFalse(shouldSkip);
+
         String image = "Image new";
         fieldDao.updateImage(field.getId(), image);
         assertEquals(image, fieldDao.getField(field.getId()).getImage());
@@ -104,6 +109,7 @@ class FieldDaoTest extends GeneralDAOTest{
 
     @Test
     void updateSport() throws SQLException, ClassNotFoundException {
+        Assumptions.assumeFalse(shouldSkip);
 
         fieldDao.updateSport(field.getId(), sport.getId());
         assertEquals(sport.getName(), fieldDao.getField(field.getId()).getSport().getName());
@@ -111,46 +117,64 @@ class FieldDaoTest extends GeneralDAOTest{
 
     @Test
     void getFieldsByFacility() throws SQLException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertFalse(fieldDao.getFieldsByFacility(facility.getId(), false).isEmpty());
     }
 
     @Test
     void getAllFields() throws SQLException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertFalse(fieldDao.getAllFields(false).isEmpty());
     }
 
     @Test
     void getFieldsByProvince() throws SQLException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertFalse(fieldDao.getFieldsByProvince(facility.getProvince()).isEmpty());
     }
 
     @Test
     void getFieldsByName() throws SQLException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertFalse(fieldDao.getFieldsByName(field.getName()).isEmpty());
     }
 
     @Test
     void getFieldsBySport() throws SQLException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertFalse(fieldDao.getFieldsBySport(field.getSport().getName()).isEmpty());
     }
 
     @Test
     void search() throws SQLException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertFalse(fieldDao.search("Campo").isEmpty());
     }
 
     @Test
     void getFieldsByOwner() throws SQLException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertFalse(fieldDao.getFieldsByOwner(facility.getOwner()).isEmpty());
     }
 
     @Test
     void reservedFields() throws SQLException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertEquals(1, fieldDao.reservedFields(reservation.getEventDate(), facility.getOwner()));
     }
 
     @Test
     void getFieldAddress() throws SQLException {
+        Assumptions.assumeFalse(shouldSkip);
+
         assertEquals(facility.getAddress(), fieldDao.getFieldAddress(field.getId()));
     }
 }
